backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
firebase-admin==6.1.0
python-dotenv==1.0.0
numpy==1.24.2
scikit-learn==1.2.2
joblib==1.2.0
pandas==1.5.3
scipy==1.10.1
gunicorn==20.1.0